Implement full basic calculator functionality.
This is the first functional and cleaned-up version of the RSE calculator. Using the command line, a user can now put in a DAQ input file of a certain format and then receive the output as an XML. Almost all basic functions have test cases for them and blocks of calculation have sample files to test the, as well. The test, physics, and engine constants are put in a YAML file that can also be passed in.